backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.2.1
uvicorn==0.24.0
better-auth==0.0.1b11
python-jose==3.3.0
PyJWT==2.10.1
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
pytz==2023.3
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
python-dateutil==2.8.2
cohere==5.5.3